refactor: embed realtime chat on dashboard
